Add files via upload
"DO NOT SELL OR MASS PRODUCE THIS IS FOR PERSONAL USE ONLY!!!"
"Commercial use, including selling or mass producing is strictly prohibited!"

Notes:
This is the first version of the Heltec V3 Mini Case. 
noted problems I'm working on are the front face clip not securely fixed though it works fine for now.
The Key hole cant hold most rings sizes and needs to be changed or adapted in V1.3 this will come.
Please Like my work and enjoy it.